+++ apps/mobile/lib/features/chat/chat_message_attachment_content.dart
@@ -57,6 +57,7 @@ final class _ChatAttachment extends ConsumerWidget {
               context,
               account: account,
               previewUri: fullScreenPreviewUri,
+              smallerPreviewUri: previewUri,
               originalUri: originalUri,
               originalContentType: mimeType,
               imageName: name,